Api v2 (#1)
* gitignore

* c# API template

* data with models for database tables

* added required and bindings between models

* first data migration and database setup

* moved entities to its own folder

* FIrst working repository

* updated repository player to async

* update to foreign keys in entities

* PlayersController

* Realised I don't need repository

* generated controllers

* the rest of controllers

* added json ignore to non neccessary relationships to avoid cycles

* added controller endpoints

* removed user from api altogether

* match overhaul

* more match adjusting

* added IsEmpty to POMR

* adjusting how crud work with references

* added one match endpoint

* added reverser property for match

* figuring out delete match (empty players working)

* player + player match delete wokring

* finally fixed the match cant delete issue in full

* Api v2 (finished pomr controller)

+++ TournamentManagerAPI/TournamentManagerAPI/Controllers/MatchesController.cs
@@ -15,10 +15,12 @@ namespace TournamentManagerAPI.Controllers
     public class MatchesController : ControllerBase
     {
         private readonly AppDBContext _context;
+        private readonly PlayerOrMatchResultsController _pomrController;
 
         public MatchesController(AppDBContext context)
         {
             _context = context;
+            _pomrController = new PlayerOrMatchResultsController(context);
         }
 
         // GET: api/Matches/5
@@ -31,6 +33,9 @@ namespace TournamentManagerAPI.Controllers
             }
             var match = await _context.Matches
                 .Include(m => m.Players)
+                .ThenInclude(p => p.Player)
+                .Include(m => m.Players)
+                .ThenInclude(p => p.Match)
                 .Include(m => m.Winner)
                 .FirstOrDefaultAsync(m => m.Id == id);
 
@@ -50,6 +55,20 @@ namespace TournamentManagerAPI.Controllers
             if (id != match.Id)
             {
                 return BadRequest();
+            }
+
+            if (match.WinnerId != null && match.PlayerRequiringResultId != null)
+            {
+                // if result was put in, updates dependent player match
+                var pomr = await _context.PlayerOrMatchResults.FindAsync((int)match.PlayerRequiringResultId);
+                if (pomr != null && !pomr.IsEmpty && !pomr.IsPlayer)
+                {
+                    pomr.IsPlayer = true;
+                    pomr.PlayerId = match.WinnerId;
+                    pomr.MatchId = null;
+                    match.PlayerRequiringResultId = null;
+                    _context.Entry(pomr).State = EntityState.Modified;
+                }
             }
 
             _context.Entry(match).State = EntityState.Modified;
@@ -82,6 +101,25 @@ namespace TournamentManagerAPI.Controllers
             {
                 return Problem("Entity set 'AppDBContext.Matches'  is null.");
             }
+
+            if(match.Players.Count > 2)
+            {
+                return BadRequest("Match cannot have more than 2 players");
+            } 
+
+            // add empty POMR (player or match result) to fill matches players
+            if(match.Players.Count < 2)
+            {
+                while(match.Players.Count < 2)
+                {
+                    match.Players.Add(new PlayerOrMatchResult()
+                    {
+                        IsEmpty = true,
+                        OriginalMatchId = match.Id
+                    });
+                }
+            }
+
             _context.Matches.Add(match);
             await _context.SaveChangesAsync();
 
@@ -96,16 +134,40 @@ namespace TournamentManagerAPI.Controllers
             {
                 return NotFound();
             }
-            var match = await _context.Matches.FindAsync(id);
+            var match = await _context.Matches
+                .Where(m => m.Id == id)
+                .Include(m => m.Players)
+                .FirstOrDefaultAsync();
+
             if (match == null)
             {
                 return NotFound();
+            }
+            
+            if (match.PlayerRequiringResultId != null)
+            {
+                return BadRequest("Cannot delete match when its result is to be used by another match.");
+            }
+            
+            match.WinnerId = null;
+            match.Winner = null;
+
+            foreach(var pomr in match.Players)
+            {
+                if (!pomr.IsEmpty) await _pomrController.EmptyPlayerOrMatchResult(pomr.Id);
             }
 
             _context.Matches.Remove(match);
             await _context.SaveChangesAsync();
 
             return NoContent();
+        }
+
+        private async Task DeleteAllPlayersOrMatchResultsWithOriginMatch(int id)
+        {
+            var players = await _context.PlayerOrMatchResults.Where(p => p.OriginalMatchId == id).ToListAsync();
+            foreach (var player in players) _context.PlayerOrMatchResults.Remove(player);
+            await _context.SaveChangesAsync();
         }
 
         private bool MatchExists(int id)
